+++ support/devcenter/generate.php
@@ -7,8 +7,7 @@ require('vendor/autoload.php');
 
 // these need updating from time to time to add new stacks and remove EOL ones
 $stacks = [
-	1 => '18', // the offset we start with here is relevant for the numbering of footnotes
-	'20',
+	1 => '20', // the offset we start with here is relevant for the numbering of footnotes
 	'22',
 ];
 // these need updating from time to time to add new series and remove EOL ones